packages/task-queue/pyproject.toml
[project]
name = "task-queue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jamie McCann", email = "jamie.mccann@umu.se" }
]
requires-python = ">=3.13"
dependencies = [
    "celery[pydantic]>=5.5.3",
    "networkx>=3.6",
    "py-fasta-validator>=0.6",
    "pydantic>=2.10.6",
    "shared",
]

# [tool.uv.workspace]
# members = ["../shared"]


[project.scripts]
task-queue = "task_queue:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "celery-types>=0.23.0",
    "docker>=7.1.0",
    "mypy>=1.15.0",
    "pika>=1.3.2",
    "pytest>=9.0.1",
    "testcontainers[rabbitmq,redis]>=4.13.3",
    "types-docker>=7.1.0.20251202",
    "types-networkx>=3.6.0.20251127",
    "types-pika-ts>=1.3.0.20250914",
]

[tool.mypy]
exclude = [
    "src/tests/.*"
]

[[tool.mypy.overrides]]
module = "testcontainers.*"
ignore_missing_imports = true
